+++ src/main/java/com/cliff/jigsaw/common/upload/ResponseForm.java
@@ -1,0 +1,2 @@
+package com.cliff.jigsaw.common.upload;public class ResponseForm {
+}
