+++ src/main/java/com/cliff/jigsaw/model/user/UserProfile.java
@@ -1,0 +1,2 @@
+package com.cliff.jigsaw.model.user;public class UserProfile {
+}
