+++ src/main/java/com/cliff/jigsaw/model/user/vo/CreateUserVo.java
@@ -1,0 +1,13 @@
+package com.cliff.jigsaw.model.vo;
+
+import lombok.Data;
+
+@Data
+public class CreateUserVo {
+    String email;
+    String phoneNumber;
+    String jobGroup;
+    String skill;
+    String field;
+    String type;
+}
