+++ src/main/java/com/cliff/jigsaw/model/user/vo/UpdateUserVo.java
@@ -1,0 +1,12 @@
+package com.cliff.jigsaw.model.vo;
+
+import lombok.Data;
+
+@Data
+public class UpdateUserVo {
+    String phoneNumber;
+    String jobGroup;
+    String skill;
+    String field;
+    String type;
+}
