+++ src/main/java/com/cliff/jigsaw/model/user/vo/UserVo.java
@@ -1,0 +1,10 @@
+package com.cliff.jigsaw.model.vo;
+
+import lombok.Data;
+
+@Data
+public class UserVo {
+
+    String email;
+
+}
