--- src/main/java/com/cliff/jigsaw/model/vo/UpdateUserVo.java
@@ -1,12 +1,0 @@
-package com.cliff.jigsaw.model.vo;
-
-import lombok.Data;
-
-@Data
-public class UpdateUserVo {
-    String phoneNumber;
-    String jobGroup;
-    String skill;
-    String field;
-    String type;
-}
