--- src/main/java/com/cliff/jigsaw/model/vo/UserVo.java
@@ -1,10 +1,0 @@
-package com.cliff.jigsaw.model.vo;
-
-import lombok.Data;
-
-@Data
-public class UserVo {
-
-    String email;
-
-}
