+++ src/main/java/com/cliff/jigsaw/repository/UserProfileRepository.java
@@ -1,0 +1,2 @@
+package com.cliff.jigsaw.repository;public class UserProfileRepository {
+}
